services/Books_backend/requirements.txt
fastapi[all]
uvicorn
gunicorn
databases
aiosqlite
python-dotenv
boto3
sentry-sdk[fastapi]

